// packages/cactus-plugin-ledger-connector-daml/src/main/typescript/generated/openapi/typescript-axios/api.ts
export interface CreateIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequest
     */
    'templateId'?: string;
    /**
     * 
     * @type {CreateIOUEndpointRequestPayload}
     * @memberof CreateIOUEndpointRequest
     */
    'payload'?: CreateIOUEndpointRequestPayload;
}
/**
 * 
 * @export
 * @interface CreateIOUEndpointRequestPayload
 */
export interface CreateIOUEndpointRequestPayload {
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'issuer'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'owner'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'currency'?: string;
    /**
     * 
     * @type {string}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'amount'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof CreateIOUEndpointRequestPayload
     */
    'observers'?: Array<string>;
}
// ...
export interface ExerciseIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'previousParticipantToken'?: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'templateId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'contractId'?: string;
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequest
     */
    'choice'?: string;
    /**
     * 
     * @type {ExerciseIOUEndpointRequestArgument}
     * @memberof ExerciseIOUEndpointRequest
     */
    'argument'?: ExerciseIOUEndpointRequestArgument;
}
/**
 * 
 * @export
 * @interface ExerciseIOUEndpointRequestArgument
 */
export interface ExerciseIOUEndpointRequestArgument {
    /**
     * 
     * @type {string}
     * @memberof ExerciseIOUEndpointRequestArgument
     */
    'newOwner'?: string;
}
// ...
export interface QueryIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof QueryIOUEndpointRequest
     */
    'participantToken'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof QueryIOUEndpointRequest
     */
    'templateIds'?: Array<string>;
    /**
     * 
     * @type {QueryIOUEndpointRequestQuery}
     * @memberof QueryIOUEndpointRequest
     */
    'query'?: QueryIOUEndpointRequestQuery;
    /**
     * 
     * @type {Array<string>}
     * @memberof QueryIOUEndpointRequest
     */
    'readers'?: Array<string>;
}
/**
 * 
 * @export
 * @interface QueryIOUEndpointRequestQuery
 */
export interface QueryIOUEndpointRequestQuery {
    /**
     * 
     * @type {number}
     * @memberof QueryIOUEndpointRequestQuery
     */
    'amount'?: number;
}
// ...
export interface QueryRawIOUEndpointRequest {
    /**
     * 
     * @type {string}
     * @memberof QueryRawIOUEndpointRequest
     */
    'participantToken'?: string;
}
// ...
export class DefaultApi extends BaseAPI {
    /**
     * 
     * @summary This method creates a simple iou countract
     * @param {CreateIOUEndpointRequest} createIOUEndpointRequest Request body for creation of iou
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public createIou(createIOUEndpointRequest: CreateIOUEndpointRequest, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).createIou(createIOUEndpointRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary This method creates a simple iou countract
     * @param {ExerciseIOUEndpointRequest} exerciseIOUEndpointRequest Request body for creation of iou
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public exerciseChoice(exerciseIOUEndpointRequest: ExerciseIOUEndpointRequest, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).exerciseChoice(exerciseIOUEndpointRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary This method starts a new instance which will get all the parties involved
     * @param {object} body Request body for getting the parties involved
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public getPartiesInvolved(body: object, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).getPartiesInvolved(body, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary This method queries DAML contracts
     * @param {QueryIOUEndpointRequest} queryIOUEndpointRequest Request body for creation of iou
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public queryContract(queryIOUEndpointRequest: QueryIOUEndpointRequest, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).queryContract(queryIOUEndpointRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary This method queries DAML contracts
     * @param {QueryRawIOUEndpointRequest} queryRawIOUEndpointRequest Request body for creation of iou
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DefaultApi
     */
    public queryRawContract(queryRawIOUEndpointRequest: QueryRawIOUEndpointRequest, options?: AxiosRequestConfig) {
        return DefaultApiFp(this.configuration).queryRawContract(queryRawIOUEndpointRequest, options).then((request) => request(this.axios, this.basePath));
    }
}